--- definitions/core/es5-1/flow_v0.104.x-/test_es5-1.js
@@ -1,12 +1,0 @@
-// @flow strict
-
-// $FlowExpectedError
-var x:string = NaN
-// $FlowExpectedError
-var y:string = Number.MAX_VALUE;
-// $FlowExpectedError
-var z:number = new TypeError().name;
-// $FlowExpectedError
-var w:string = parseInt("...");
-// $FlowExpectedError
-const ra: $ReadOnlyArray<number> = Object.freeze({q:2});
