+++ definitions/core/es5-1/flow_v0.104.x-v0.150.x/test_es5-1.js
@@ -1,0 +1,12 @@
+// @flow strict
+
+// $FlowExpectedError[incompatible-type]
+var x:string = NaN
+// $FlowExpectedError[incompatible-type]
+var y:string = Number.MAX_VALUE;
+// $FlowExpectedError[incompatible-type]
+var z:number = new TypeError().name;
+// $FlowExpectedError[incompatible-type]
+var w:string = parseInt("...");
+// $FlowExpectedError[incompatible-type]
+const ra: $ReadOnlyArray<number> = Object.freeze({q:2});
