+++ definitions/core/jsx/flow_v0.83.x-/test_jsx.js
@@ -1,4 +1,5 @@
 // @flow
+import { describe, it } from 'flow-typed-test';
 import * as React from 'react';
 
 describe('jsx', () => {
